Fix cached node handling in social network adapter
Previously, cached nodes were not pushed to the node stack correctly,
resulting in incorrect behaviour.  This commit also ensures that when
cached entries for a cached node are used as the model data, they are
filtered according to the current filter set, in case the filters have
changed since the cache entries were retrieved.