[nemo-qml-plugin-social] Ensure Facebook data cache integrity
Previously, cache entries for the wrong node could be inserted or
removed erroneously, which could lead to a crash.  This commit ensures
that any outstanding network requests are deleted if the client sets
the current node to something different, which prevents the cache from
being populated with incorrect data.

It also ensures that the cache-index built from instantiated items is
populated and depopulated correctly, to avoid spurious object creation.