Fixes to name sorter
This commit first renamed NameSorterInterface
to AlphabeticalSorterInterface. That interface
now uses a property, "field" to determine which
field should be used to perfor the sort.

Some behaviour are restored: the abstract base
sorter perform a sort based on type, and the
alphabetical sorter also sorts based on type before
using the provided "field".