Support for loading next and previous pages of results
This commit implements a paging system, allowing previous
pages of data and next pages to be loaded via the loadPrevious()
and loadNext() method in SNI.

It also includes modifications that allows the model to prepend
and append data to the different nodes. Facebook is fully
supported.

It finally includes fixes in the sorting method, so that the
data that is displayed is automatically resorted when sorters
are changed.