[feature] Multimodel changes
This commit introduces the multimodel API, that separates
the SocialNetworkInterface into two classes. One, still
called SocialNetworkInterface, is used to makage the cache
and interfaces with given social networks. The other class,
SocialNetworkModelInterface, is used to carry data to be
displayed. It uses the cache in SocialNetworkInterface to
get these data.