[bugfix] Bugfixes to SNI and FI
This commit includes urgent bugfixes to SNI and FI, that caused
the current node not being refreshed, and also causes error when
the filters were empty.

Also profit from that commit to kills .moc and .obj folders for storing
mocs and binaries. It is better for shadow builds on desktop.