Cleanups, refactors and bugfixes
This commit includes a small redesign of the relation between
SNI and SNMI. Node classes do not contain the model they are
representing, nor the QNetworkRequest that they are using
to get data, therefore respecting their role of "pure data" class.

Models and replies are now stored by SNIp. Updates are still
done by SNIp as well.

This commit also includes the feature to reload a node, that can
be done via SNMI::repopulate().

Some bugs, like ICII not being able to be populated, or UserPicture
not being loaded, are fixed, and there are tests in the Facebook
social test.