[bugfix] Splitted addModel into adding and populating
SNIp only had a "addModel" method that was used to add a model
and populate the Node associated to the model. Now the two
operations are splitted. The model will only call addModel when
the socialNetwork is set. Populate is called when SNMI::populate()
or SNMI::repopulate() is called.

This commit also fix a bug that might happen, when, when the user
uses the same model to load a different node, the old node
was not purged.

It also cleaned two methods that were not removed during the
refactor in FacebookInterface.