[bugfix] Be able to retrieve number of likes and comments
This commit brings (back the capability to retrieve the
number of likes and comments that was posted in a Post, Album or
Photo. It was removed because of a Facbeook API change.

It also fixes some bugs, notably the one making picture of
"me" impossible to load in the demo application. This fix
is done by introducing the "alias" map in SNIp. Comparison
between nodes and models takes aliases of the identifier
in consideration (like the one between "me" and the actual
identifier in Facebook).

The UserPicture loading is also being cleaner, and now
use related data method to load, making reloading of only
this component possible (fixing another bug).

The last bug that was fixed is the capability to store the
type that was guessed, and do not have to guess it again.

It also fixes the small warning that happens when running qmake
in the test folder in Qt5 (because the qt5 folder often exists).