Handle the case when the SNI is destroyed first
If a SNI is destroyed first, it can often lead to segfaults,
since it destroys items, and the models destroy them again
afterwards.

The fix consist of telling to all models to clear themselves,
in order not to have any item used by the models before deleting
them, starting with the Node, and the with the CacheEntries.