Fix for page handling
Page handling was partly broken before, since, when
pushing data after or before in a SNMI caused information
loss about pages.

Specifically, if we consider data that have previous
and next data, we store the previous cursors and next
cursors in Node::extraInfo.

However, when newer data got pushed (for example appended),
newer cursors for _both_ next and previous got inserted
in Node::extraInfo, so we lost information (in this example,
the cursors for prepending data).

This caused problems for navigation if a data stream supports
both loading next and previous pages.