[feature] Workaround for removing retweets
The workaround consists of getting the latest tweets someone
made and then search for the retweet inside them. If it cannot
be found, there will be a failure in removing the RT.

After removing a RT, there might be some error while retweeting
again right after (maybe because Twitter think that you still
have the tweet in the timeline, and prevent you to re-retweet
directly). It is something rather unpredictible and impossible
to prevent.