[bugfix] Prevent crash if SNI is destroyed and SNMI being destroyed too
This commit prevents a crash if SNI is destroyed where SNMI is
destroyed too. If we call QALM::endRemoveRows() while the view
is being destroyed, it might lead to some double free (item models
are being removed first by the view and again by the model). If we
don't clean when deleting the SNI, we cover most user-case and
prevent this crash.

The only issue that can happen is that we keep the SNMI and
destroys the SNI. Then, data in the SNMI would stay. It is
not really nice, but shold be so rare that we can discard
this case.